Basic bug fixed - but still WIP
Design flaw: inference of NpgsqlDbType and DbType currently consults the global type handler only.
This means that any mappings (composite, enum or otherwise) done on a specific connection
aren't taken into account (see test TestEnumType which is currently broken).

When inferring in NpgsqlParameter, see if the parameter is on a command with a connection, and if so,
infer via its own type mapper.