remove-command-locking
If this helps performance (and it will), we can still have a variable
guard which we check to try and help users locate concurrency
issues (maybe even Interlocked). Or instead of another variable,
just use the existing state variable....

We will probably still need the SemaphoreSlim for when keepalive
is turned on though.