Aborted work on command caching
Aborted work on caching entire commands, rather than individual
statements. The idea was to avoid the SQL parsing that has to
happen every time, to break down CommandText into statements before
looking if there are cached prepared statements for them.

Some basic benchmarks showed around 3% perf improvement for
multi-statement commands (10 statements). The downside was considerable
complexity as commands have to be tracked as well as statements. Both
statements and commands need to have reference count.
Also, a maximum number of *statements* still has to be tracked (that's
the expensive resource), but to free up statements we need to deallocate
commands, which themselves references other statements possible in use
by other commands...

This was ultimately dropped because this only makes sense because of
the lack of a proper batching API in ADO.NET. In other words, if the
user could provide statements in a sane way (not concatenated inside
CommandText), this would become meaningless.

One good thing that came out of this: merged NpgsqlStatement and
PreparedStatement.

See #1701