WIP on resetting type mapper state efficiently
However... Going this way doesn't seem to work well with super-short-lived
connections, which is what we'll have if we do multiplexing. It would mean
that on each and every execution we modify the type mapper, execute, and
then roll back. It may be OK, or may not.

Maybe the fundamental problem is that we allow too much flexibility: people
don't typically need per-connection mapping/binding, they need per-database
mapping/binding.

Maybe it's time to reconsider our model, and replace connection-specific
with database-specific. The problem is that bindings are about type handlers,
and type handlers are currently connector-specific.